pyproject.toml
[project]
name = "e-commerce-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.5.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "lxml>=6.0.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "selenium>=4.41.0",
    "semantic-router>=0.1.12",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
]
